requirements.txt
fastapi~=0.115.12
uvicorn
sqlalchemy~=2.0.41
psycopg2-binary
geoalchemy2
websockets
pydantic~=2.11.5
python-dotenv~=1.1.0
pytest~=8.3.5
requests
shapely
pyproj
pillow
httpx
postgis
fastmcp
pytest-asyncio
websocket-client
pydantic-settings~=2.9.1
pytest-cov 
bcrypt
python-jose
pydantic[email]
SimpleLLMFunc